Move helper methods from the capa tests module __init__ file
