+++ common/lib/capa/capa/tests/test_shuffle.py
@@ -3,7 +3,7 @@
 import unittest
 import textwrap
 
-from . import test_capa_system, new_loncapa_problem
+from capa.tests.helpers import test_capa_system, new_loncapa_problem
 from capa.responsetypes import LoncapaProblemError
