+++ common/lib/capa/capa/tests/test_util.py
@@ -4,7 +4,7 @@ Tests capa util
 import unittest
 from lxml import etree
 
-from . import test_capa_system
+from capa.tests.helpers import test_capa_system
 from capa.util import compare_with_tolerance, sanitize_html, get_inner_html_from_xpath
